+++ apps/modernjs-ssr/dynamic-nested-remote/modern.config.ts
@@ -4,19 +4,16 @@ import { moduleFederationPlugin } from '@module-federation/modern-js';
 // https://modernjs.dev/en/configure/app/usage
 export default defineConfig({
   dev: {
-    port: 3009,
+    port: 3054,
   },
   runtime: {
     router: true,
-  },
-  output: {
-    disableTsChecker: true,
   },
   server: {
     ssr: {
       mode: 'stream',
     },
-    port: 3009,
+    port: 3054,
   },
   plugins: [appTools(), moduleFederationPlugin()],
 });