+++ apps/modernjs-ssr/dynamic-remote-new-version/modern.config.ts
@@ -4,7 +4,7 @@ import { moduleFederationPlugin } from '@module-federation/modern-js';
 // https://modernjs.dev/en/configure/app/usage
 export default defineConfig({
   dev: {
-    port: 3011,
+    port: 3056,
   },
   runtime: {
     router: true,
@@ -14,9 +14,7 @@ export default defineConfig({
       mode: 'stream',
     },
   },
-  output: {
-    disableTsChecker: true,
-  },
+
   tools: {
     webpack: (config) => {
       config.output!.uniqueName = 'modernjs-ssr-dynamic-remote-new-version';