+++ apps/modernjs-ssr/host/module-federation.config.ts
@@ -2,8 +2,8 @@ import { createModuleFederationConfig } from '@module-federation/modern-js';
 export default createModuleFederationConfig({
   name: 'host',
   remotes: {
-    remote: 'remote@http://localhost:3006/mf-manifest.json',
-    nested_remote: 'nested_remote@http://localhost:3007/mf-manifest.json',
+    remote: 'remote@http://localhost:3051/mf-manifest.json',
+    nested_remote: 'nested_remote@http://localhost:3052/mf-manifest.json',
   },
   shared: {
     react: { singleton: true },