+++ apps/modernjs-ssr/host/src/routes/all/page.tsx
@@ -4,11 +4,11 @@ import { loadRemote, registerRemotes } from '@modern-js/runtime/mf';
 registerRemotes([
   {
     name: 'dynamic_nested_remote',
-    entry: 'http://localhost:3009/mf-manifest.json',
+    entry: 'http://localhost:3054/mf-manifest.json',
   },
   {
     name: 'dynamic_remote',
-    entry: 'http://localhost:3008/mf-manifest.json',
+    entry: 'http://localhost:3053/mf-manifest.json',
   },
 ]);
 
@@ -26,7 +26,7 @@ const DynamicRemote = React.lazy(() =>
 
 const Remote = React.lazy(() => {
   return import('remote/Image').then((r) => {
-    return new Promise((resolve) => {
+    return new Promise<typeof import('remote/Image')>((resolve) => {
       setTimeout(() => {
         resolve(r);
       }, 2000);
@@ -36,7 +36,7 @@ const Remote = React.lazy(() => {
 
 const NestedRemote = React.lazy(() => {
   return import('nested_remote/Content').then((r) => {
-    return new Promise((resolve) => {
+    return new Promise<typeof import('nested_remote/Content')>((resolve) => {
       setTimeout(() => {
         resolve(r);
       }, 1000);
@@ -61,7 +61,7 @@ const Index = (): JSX.Element => {
           {/* remote */}
           <tr>
             <td>✅</td>
-            <td>This component is from a remote(localhost:3006)</td>
+            <td>This component is from a remote(localhost:3051)</td>
             <td>
               <button
                 style={{ marginBottom: '1rem' }}
@@ -71,7 +71,7 @@ const Index = (): JSX.Element => {
               </button>
             </td>
             <td>
-              <Suspense loading={'loading remote for 2000ms'}>
+              <Suspense fallback={'loading remote for 2000ms'}>
                 <Remote />
               </Suspense>
             </td>
@@ -80,7 +80,7 @@ const Index = (): JSX.Element => {
           <tr>
             <td>✅</td>
             <td>
-              This component is from a remote(localhost:3007) which nest a
+              This component is from a remote(localhost:3052) which nest a
               static remote
             </td>
             <td>
@@ -92,7 +92,7 @@ const Index = (): JSX.Element => {
               </button>
             </td>
             <td>
-              <Suspense loading={'loading nested remote for 1000ms'}>
+              <Suspense fallback={'loading nested remote for 1000ms'}>
                 <NestedRemote />
               </Suspense>
             </td>
@@ -100,7 +100,7 @@ const Index = (): JSX.Element => {
           {/* dynamic remote */}
           <tr>
             <td>✅</td>
-            <td>This component is from a dynamic remote(localhost:3008)</td>
+            <td>This component is from a dynamic remote(localhost:3053)</td>
             <td>
               <button
                 style={{ marginBottom: '1rem' }}
@@ -110,7 +110,7 @@ const Index = (): JSX.Element => {
               </button>
             </td>
             <td>
-              <DynamicRemote />
+              <DynamicRemote text="" />
             </td>
           </tr>
 
@@ -118,7 +118,7 @@ const Index = (): JSX.Element => {
           <tr>
             <td>✅</td>
             <td>
-              This component is from a dynamic remote(localhost:3009) which nest
+              This component is from a dynamic remote(localhost:3054) which nest
               a dynamic remote
             </td>
             <td>