+++ apps/modernjs-ssr/nested-remote/modern.config.ts
@@ -4,22 +4,16 @@ import { moduleFederationPlugin } from '@module-federation/modern-js';
 // https://modernjs.dev/en/configure/app/usage
 export default defineConfig({
   dev: {
-    port: 3007,
+    port: 3052,
   },
   runtime: {
     router: true,
-  },
-  output: {
-    disableTsChecker: true,
-  },
-  source: {
-    enableAsyncEntry: true,
   },
   server: {
     ssr: {
       mode: 'stream',
     },
-    port: 3007,
+    port: 3052,
   },
   plugins: [
     appTools({