+++ apps/node-remote/webpack.config.js
@@ -29,6 +29,5 @@ module.exports = composePlugins(withNx(), (config) => {
       },
     }),
   );
-  debugger;
   return config;
 });