+++ packages/sdk/rollup.config.js
@@ -6,6 +6,12 @@ module.exports = (rollupConfig, _projectOptions) => {
       targets: [{ src: 'packages/sdk/LICENSE', dest: 'packages/sdk/dist' }],
     }),
   );
+  // rollupConfig.plugins.push({
+  //   name: 'custom-dynamic-import',
+  //   renderDynamicImport({ moduleId }) {
+  //     return { left: 'import(', right: ')' };
+  //   },
+  // });
 
   return rollupConfig;
 };