+++ apps/modernjs/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @module-federation/modernjs
 
+## 0.1.33
+
+### Patch Changes
+
+- Updated dependencies [4488064]
+  - @module-federation/enhanced@0.2.6
+
+## 0.1.32
+
+### Patch Changes
+
+- @module-federation/enhanced@0.2.5
+
+## 0.1.31
+
+### Patch Changes
+
+- @module-federation/enhanced@0.2.4
+
 ## 0.1.30
 
 ### Patch Changes