+++ apps/router-demo/router-remote2-2002/CHANGELOG.md
@@ -1,5 +1,27 @@
 # remote2
 
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies [4488064]
+  - @module-federation/enhanced@0.2.6
+  - @module-federation/bridge-react@0.2.6
+
+## 1.0.6
+
+### Patch Changes
+
+- @module-federation/enhanced@0.2.5
+- @module-federation/bridge-react@0.2.5
+
+## 1.0.5
+
+### Patch Changes
+
+- @module-federation/enhanced@0.2.4
+- @module-federation/bridge-react@0.2.4
+
 ## 1.0.4
 
 ### Patch Changes