+++ packages/bridge/bridge-react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @module-federation/bridge-react
 
+## 0.2.6
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.2.6
+
+## 0.2.5
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.2.5
+
+## 0.2.4
+
+### Patch Changes
+
+- @module-federation/bridge-shared@0.2.4
+
 ## 0.2.3
 
 ### Patch Changes