+++ packages/managers/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @module-federation/managers
 
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies [91bf689]
+  - @module-federation/sdk@0.2.6
+
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies [8cce571]
+  - @module-federation/sdk@0.2.5
+
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [09b792d]
+- Updated dependencies [09b792d]
+  - @module-federation/sdk@0.2.4
+
 ## 0.2.3
 
 ### Patch Changes