+++ packages/runtime-tools/CHANGELOG.md
@@ -1,5 +1,29 @@
 ## [1.0.1-canary.1](https://github.com/module-federation/core/compare/runtime-1.0.0...runtime-1.0.1-canary.1) (2023-12-06)
 
+## 0.2.6
+
+### Patch Changes
+
+- @module-federation/runtime@0.2.6
+- @module-federation/webpack-bundler-runtime@0.2.6
+
+## 0.2.5
+
+### Patch Changes
+
+- Updated dependencies [8cce571]
+  - @module-federation/runtime@0.2.5
+  - @module-federation/webpack-bundler-runtime@0.2.5
+
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies [60cd259]
+- Updated dependencies [09b792d]
+  - @module-federation/runtime@0.2.4
+  - @module-federation/webpack-bundler-runtime@0.2.4
+
 ## 0.2.3
 
 ### Patch Changes