+++ packages/third-party-dts-extractor/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @module-federation/third-party-dts-extractor
 
+## 0.2.6
+
+## 0.2.5
+
+## 0.2.4
+
 ## 0.2.3
 
 ## 0.2.2